pyproject.toml
[tool.poetry]
name = "pynam"
version = "0.1.0"
description = ""
authors = ["Vernon Swanepoel <frantic.naturalist@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"
Flask = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
